+++ shared/src/pretty_print.rs
@@ -7,17 +7,14 @@ pub const YELLOW: &str = "\x1B[33m";
 pub const ORANGE: &str = "\x1B[38;5;208m";
 pub const LIGHT_GRAY: &str = "\x1B[90m";
 
-#[inline(always)]
 pub fn print_success(msg: impl Display) {
     println!("{GREEN}[+]{RESET} {msg}");
 }
 
-#[inline(always)]
 pub fn print_info(msg: impl Display) {
     println!("{YELLOW}[i]{RESET} {msg}");
 }
 
-#[inline(always)]
 pub fn print_failed(msg: impl Display) {
     println!("{RED}[-]{RESET} {msg}");
 }