+++ RELEASE_NOTES.md
@@ -6,6 +6,20 @@ pulling updates.
 
 **IN ANY CASE ALWAYS BACKUP YOUR PROFILES BEFORE UPDATING!!!!**
 
+## &#128679; v 0.5
+
+### &#128679; Breaking changes
+
+- Introduced the .svc binary which builds as part of your build package from the C2. There is a new required field in the profile, which is **svc_name**. Read more in the Wyrm Docs profile section as to how to use this field. In short, the value of this field (required) is passed to the Windows Service Control Manager when the service binary is run.
+
+### Non-breaking changes
+
+- Introduced the **string scrubber**!
+  - The string scrubber automatically scrubs 'implant.dll' from the export name of the Wyrm DLL.
+  - The string scrubber allows through a malleable profile the ability to scrub certain strings from the binary. **Warning:** this interprets bytes like for like and either allows you to replace them, or zero them out. This could lead to accidental pattern collisions with machine code / other artifacts, so if you are using this feature, be sure to test the binary before deployment on a red team op!
+- Added download counter for staged resources (visible in new log file, and on the staged resources GUI page).
+- Fixed bug (again..) that was preventing messages showing in the GUI, even though they were processed by the client. Hopefully that is the end of that bug!
+
 ## v 0.4.4
 
 - Introduces the profile options to build custom DLL export names, as well as define custom machine code to run at an export. This could be used for DLL Sideloading (better support for that coming later, but it should still work in some cases), OPSEC, or just causing a bit of mayhem for a blue teamer.